1：工程命名整理 2：CodeGenerater bug修复

+++ CodeGenerater/src/main/java/my/CodeGenerater/CodeWriter.java
@@ -3,6 +3,8 @@ package my.CodeGenerater;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Paths;
 
 public class CodeWriter {
 	private static String default_suffix = ".java";
@@ -22,10 +24,40 @@ public class CodeWriter {
 			dir.mkdirs();
 
 		File file = new File(absPath + "/" + fileName + (suffix == null ? default_suffix : suffix));
-		FileWriter fw = new FileWriter(file);
-		fw.write(content);
-		fw.flush();
-		fw.close();
+		StringBuffer sb = new StringBuffer();
+		if(file.exists()){
+			System.out.println(file.getPath() +" 已经存在，是否覆盖？yes/no");
+			//至少5个字节，yes/no + 13 10
+			//yes = [ x x x ,13,10]
+			//no = [x x,13,10,0]
+			byte[] renew = new byte[5];
+			System.in.read(renew);
+			
+			sb.append(new String(renew));
+			String input = sb.toString().toLowerCase().trim();
+			if("yes".equals(input)){
+				FileWriter fw = new FileWriter(file);
+				fw.write(new String(content.getBytes("utf-8")));
+				fw.flush();
+				fw.close();
+			}else if("no".equals(input)){
+				File dest = new File(absPath + "/" + fileName + (suffix == null ? default_suffix : suffix)+".bak."+System.currentTimeMillis());
+				file.renameTo(dest);
+				FileWriter fw = new FileWriter(file);
+				fw.write(new String(content.getBytes("utf-8")));
+				fw.flush();
+				fw.close();
+			}else{
+				System.out.println("输入错误");
+			}
+		}else{
+			FileWriter fw = new FileWriter(file);
+			fw.write(new String(content.getBytes("utf-8")));
+			fw.flush();
+			fw.close();
+		}
+		
+		
 	}
 
 	public void write(String content) throws IOException {
